Fixing a delete Tier problem
When deleting a tier, character might get shifted down or disappear completely from the list. This is due to how we store ranking as an array and when we don't change the hero's rating values it either goes off the bound causing it to disappear or it gets moved to an improper tier.

For example, if Alfonse's rating is stored as 7, in a tier list with 8 rankings. If we delete one of the tier, this will cause alfonse to disappear. Since the largest index for the new tier is 6 because the tier list is now size 7.

This bug will not affect anyone below the deleted tier so we only need to fix any hero above the tierIndex.